reorganize app structure
